% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlastfiles.R
\name{findlastfiles}
\alias{findlastfiles}
\title{Returns the location of the most recent peptides and modified conditions file}
\usage{
findlastfiles(data)
}
\arguments{
\item{data}{pep_data generated by preppetide, it is a list where 4th element contains information about conditions}
}
\value{
path for the last folder, peptides file and conditions file
}
\description{
Returns the location of the most recent peptides and modified conditions file
}
\examples{
findlastfiles()
}
