% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlastpepfile.R
\name{findlastpepfile}
\alias{findlastpepfile}
\title{returns the location of the most recent peptides and modified conditions file}
\usage{
findlastpepfile()
}
\value{
last folder
}
\description{
returns the location of the most recent peptides and modified conditions file
}
\examples{
findlastpepfile()
}
