% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlastpepfile_v2.R
\name{findlastpepfile_v2}
\alias{findlastpepfile_v2}
\title{Returns the location of the most recent peptides and modified conditions file (v2)}
\usage{
findlastpepfile_v2(data)
}
\arguments{
\item{data}{pep_data generated by preppetide, it is a list where 4th element contains information about conditions}
}
\value{
path for the last folder, peptides file and conditions file
}
\description{
Returns the location of the most recent peptides and modified conditions file (v2)
}
\examples{
findlastpepfile()
}
