% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepquantify_funs.R
\name{pepquantify_funs}
\alias{pepquantify_funs}
\title{Nested functions of the pepquantify package}
\usage{
pepquantify_funs(
  data,
  imputation = FALSE,
  n_element_peptide = 1,
  condition1,
  condition2,
  n_condition_1 = 3,
  n_condition_2 = 3,
  min_pep = 2,
  downshift = 1.8,
  width = 0.3,
  n_ko_like = 2,
  fraction_valid = 1,
  second_condition = 1,
  seed = 1234
)
}
\arguments{
\item{data}{list of two containing peptide and protein group data generated by the read functions of the pepquant package}

\item{imputation}{if true imputation is performed if false not}

\item{n_element_peptide}{peptide data is nth element}

\item{condition1}{name of the first condition that should be compared (note that order matters for the fold-change direction)}

\item{condition2}{name of the second condition that should be compared (note that order matters for the fold-change direction)}

\item{n_condition_1}{minimum number of the valid values in the first condition (this value should be minimum two but default pepquant value is three)}

\item{n_condition_2}{minimum number of the valid values in the second condition (this value should be minimum two but default pepquant value is three)}

\item{min_pep}{minimum number of peptides}

\item{downshift}{see the perseus documentation "Replace missing values from normal distribution"}

\item{width}{see the perseus documentation "Replace missing values from normal distribution"}

\item{n_ko_like}{minimum number of peptides that should have missing and valid value behaviour (max in one less than 2 in second or otherwise by user defined criteria) default 2}

\item{fraction_valid}{between 0-1. 1 means that imputed peptides are taken into account if they are present in all samples of one of the conditions, 0.5 means if they are present in the half of the samples of one of the conditions. default 1}

\item{second_condition}{maximum acceptable number of valid values in other condition when fraction valid is met in the other default 1}

\item{seed}{set seed as values for imputation are derived randomly, seed makes sure the reproducibility. default 1234}
}
\value{
locations_last
}
\description{
Nested functions of the pepquantify package
}
\examples{
pepquantify_funs(data, condition1, condition2)
}
