% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptimpute.R
\name{peptimpute}
\alias{peptimpute}
\title{Performs missing value imputation (Perseus-like approach)}
\usage{
peptimpute(
  data,
  downshift = 1.8,
  width = 0.3,
  n_ko_like = 2,
  fraction_valid = 1,
  second_condition = 1,
  seed = 1234
)
}
\arguments{
\item{data}{this is the data that is returned from the preppetide function}

\item{downshift}{see the Perseus documentation "Replace missing values from normal distribution"}

\item{width}{see the Perseus documentation "Replace missing values from normal distribution"}

\item{n_ko_like}{minimum number of peptides that should have missing and valid value pattern (all valid in one condition, less than 2 in the second or otherwise by user defined criteria) default 2}

\item{fraction_valid}{between 0-1. 1 means that imputed peptides are taken into account if they are present in all samples of one of the conditions, 0.5 means if they are present in the half of the samples of one of the conditions. default 1}

\item{second_condition}{maximum acceptable number of valid values in other condition when fraction valid is met in the other, default 1}

\item{seed}{set seed as values for imputation are derived randomly, seed makes sure the reproducibility. default 1234}
}
\description{
By default MS-EmpiRe does not require imputation, but requires at least two valid values in each condition.
In extreme cases it could be that the peptide is consistently detected in one of the conditions but is not found in the second condition,
such peptides might be interesting to futher explore or even visualize on the volcano plot. This function founds such peptides and imputes missing values. However after quantitative analysis generated p-values and fold-changes must be taken into
account with caution. This function tries to take those imputed peptides for quantification that are chosen with very stringend criterias. Additionally, if protein already have
peptides more than min_pep (see the preppetide function), no imputation will be performed even if there are peptides for this protein that meet imputation criteria.
For the criterias see the options of this function.
}
\examples{
peptimpute(data)
}
