% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preppeptide.R
\name{preppeptide}
\alias{preppeptide}
\title{Extract the necessary information for the peptide data}
\usage{
preppeptide(
  data,
  n_element_peptide = 1,
  condition1 = "P",
  condition2 = "wt",
  n_condition_1 = 3,
  n_condition_2 = 3,
  min_pep = 2
)
}
\arguments{
\item{data}{list of two containing peptide and protein group data generated by the read functions of the pepquant package}

\item{n_element_peptide}{peptide data is the nth element (usually not necessary to change)}

\item{condition1}{name of the first condition that should be compared (note that order matters for the fold-change direction)}

\item{condition2}{name of the second condition that should be compared (note that order matters for the fold-change direction)}

\item{n_condition_1}{minimum number of the valid values in the first condition (this value should be at least two, but default pepquant value is three)}

\item{n_condition_2}{minimum number of the valid values in the second condition (this value should be at least two, but default pepquant value is three)}

\item{min_pep}{minimum number of peptides for each protein}
}
\value{
filtered_data_dual which is the list of 6. This output should always be generated but is only useful if imputation is performed. Also, conditions and peptide file (without imputation) will be written in the designated folder.
}
\description{
This function keeps only the necessary columns that are needed for the next steps. Importantly, dataset will be reduced to a dual comparisons by passing proper group names in the condition1 and condition2.
This is important since MS-EmpiRe in one analysis can quantify dual comparisons only. Also, MS-EmpiRe requires minimum two valid values in EACH condition, so in this step filtering for valid values
is also performed by user defined settings (see also imputation step of the pepquant package)
}
\examples{
preppeptide(data, condition1, condition2)
}
