% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultstidy.R
\name{resultstidy}
\alias{resultstidy}
\title{Tidy the results generated by MS-EmpiRe}
\usage{
resultstidy(fc_threshold = 1.5, data, data2)
}
\arguments{
\item{fc_threshold}{minimum fold change for the protein to be considered differentially abundant (in natural scale) defualt 1.5}

\item{data}{dataset retured from the read functions of the pepquantify package}

\item{data2}{this is the output of the pepquantify package read functions}
}
\value{
msempire_results
}
\description{
Tidy the results generated by MS-EmpiRe
}
\examples{
resultstidy(fc_threshold = 1.5, data)
}
