% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlastpepfile.R
\name{findlastpepfile}
\alias{findlastpepfile}
\title{Returns the location of the most recent peptides and modified conditions file}
\usage{
findlastpepfile()
}
\value{
path for the last folder, peptides file and conditions file
}
\description{
Returns the location of the most recent peptides and modified conditions file
}
\examples{
findlastpepfile()
}
